Directly unmarshal JSON received from serial port
Old parsing code is removed. Arduino sends JSON directly to serial
port and the collector unmarshal the JSON and package it into a
request for server.

+++ collector/collector/request.go
@@ -12,6 +12,6 @@ type signalPublishRequest struct {
 
 // markSpacePair represents each bit for the frame
 type markSpacePair struct {
-	Mark  float64 `json:"mark"`
-	Space float64 `json:"space"`
+	Mark  int `json:"mark"`
+	Space int `json:"space"`
 }
